+++ bsp/bouffalo_lab/libraries/bl_mcu_sdk/drivers/lhal/src/bflb_gpio.c
@@ -4,8 +4,8 @@
 void bflb_gpio_init(struct bflb_device_s *dev, uint8_t pin, uint32_t cfgset)
 {
     uint32_t cfg = 0;
-    uint32_t reg_base;
-    uint32_t cfg_address;
+    size_t reg_base;
+    size_t cfg_address;
     uint8_t drive;
     uint8_t function;
     uint16_t mode;
@@ -202,7 +202,7 @@ uint32_t bflb_gpio_pin32_63_read(struct bflb_device_s *dev)
 
 void bflb_gpio_int_init(struct bflb_device_s *dev, uint8_t pin, uint8_t trig_mode)
 {
-    uint32_t reg_base;
+    size_t reg_base;
     uint32_t cfg_address;
     uint32_t regval;
 
@@ -234,7 +234,7 @@ void bflb_gpio_int_init(struct bflb_device_s *dev, uint8_t pin, uint8_t trig_mod
 
 void bflb_gpio_int_mask(struct bflb_device_s *dev, uint8_t pin, bool mask)
 {
-    uint32_t reg_base;
+    size_t reg_base;
     uint32_t cfg_address;
     uint32_t regval;
 
@@ -272,7 +272,7 @@ bool bflb_gpio_get_intstatus(struct bflb_device_s *dev, uint8_t pin)
 
 void bflb_gpio_int_clear(struct bflb_device_s *dev, uint8_t pin)
 {
-    uint32_t reg_base;
+    size_t reg_base;
     uint32_t cfg_address;
     uint32_t regval;
 
@@ -298,7 +298,7 @@ void bflb_gpio_int_clear(struct bflb_device_s *dev, uint8_t pin)
 
 void bflb_gpio_uart_init(struct bflb_device_s *dev, uint8_t pin, uint8_t uart_func)
 {
-    uint32_t reg_base;
+    size_t reg_base;
     uint32_t regval;
     uint8_t sig;
     uint8_t sig_pos;
@@ -412,7 +412,7 @@ void bflb_gpio_uart_init(struct bflb_device_s *dev, uint8_t pin, uint8_t uart_fu
 int bflb_gpio_feature_control(struct bflb_device_s *dev, int cmd, size_t arg)
 {
     int ret = 0;
-    uint32_t reg_base;
+    size_t reg_base;
     uint32_t regval;
     uint8_t pin = arg;
